+++ R/frag.graph.reduce.R
@@ -1,28 +1,28 @@
 
 
 .reduce.without.conservation <- function(g, final.frag.nr) {
-  # delete the exact number of vertices, with no warranty to conserve the number of connected components
+  # delete the exact number of vertices, with no insurance to preserve the number of connected components
   selected.vertice <- sample(igraph::V(g), 1)
   g2 <- igraph::delete_vertices(g, selected.vertice)
-  g2 <-
-    igraph::delete_vertices(g2, igraph::V(g2)[igraph::degree(g2) == 0])
-  if (!igraph::gorder(g2) < final.frag.nr) {
+  g2 <- igraph::delete_vertices(g2, igraph::V(g2)[igraph::degree(g2) == 0]) # delete singletons
+  if (igraph::gorder(g2) >= final.frag.nr) {
     g <- igraph::delete_vertices(g, selected.vertice)
-    g <-
-      igraph::delete_vertices(g, igraph::V(g)[igraph::degree(g) == 0])
-  }
+    g <- igraph::delete_vertices(g, igraph::V(g)[igraph::degree(g) == 0])
+  } else{ g$is.reducible <- FALSE }
   g
 }
 
 .reduce.with.conservation <- function(g) {
-  # delete as much vertices as possible while no decreasing the number of connected components
-  igraph::V(g)$component.id <- igraph::clusters(g)$membership
+  # delete a vertice as far as it does not decrease the number of connected components
+  igraph::V(g)$component.id <- igraph::components(g)$membership
   noDyad <- names(table(igraph::V(g)$component.id)[table(igraph::V(g)$component.id) > 2])
-  frag.selection <- (igraph::V(g)$component.id %in% noDyad) & (igraph::degree(g) == 1)
-  if (sum(frag.selection) > 0) {
+  # exclude vertices in dyadic components and articulations points:
+  frag.selection <- (igraph::V(g)$component.id %in% noDyad) & ( ! V(g) %in% igraph::articulation_points(g) )
+  
+  if (sum(frag.selection)) {
     selected.frag.to.delete <- sample(igraph::V(g)[frag.selection], 1)
     g <- igraph::delete_vertices(g, selected.frag.to.delete)
-  }
+  } else{ g$is.reducible <- FALSE }
   g
 }
 
@@ -32,21 +32,23 @@ frag.graph.reduce <- function(graph = NULL, n.frag.to.remove = NULL, conserve.ob
   if( ! is.logical(conserve.objects.nr)){stop("'conserve.objects.nr' must be TRUE or FALSE.")}
   if( ! is.numeric(n.frag.to.remove)){stop("'n.frag.to.remove' must be an integer value.")}
   
+  graph$is.reducible <- TRUE
+  
   final.frag.nr <- igraph::gorder(graph) - n.frag.to.remove
   
   if (conserve.objects.nr) {
     is.reducible <- TRUE
-    while ((igraph::gorder(graph) > final.frag.nr) & is.reducible) {
-      size.save <- igraph::gorder(graph)
+    while ((igraph::gorder(graph) > final.frag.nr) & graph$is.reducible) {
       graph <- .reduce.with.conservation(graph)
-      if(igraph::gorder(graph) == size.save){is.reducible <- FALSE}
     }
   } else {
-    while (igraph::gorder(graph) > final.frag.nr) {
+    while ((igraph::gorder(graph) > final.frag.nr) & graph$is.reducible) {
       graph <- .reduce.without.conservation(graph, final.frag.nr)
     }
   }
-  graph
+  igraph::delete_graph_attr(graph, "is.reducible")
 }
 
 
+
+